Move backend selection logic to child classes
